requirements.txt
# CLI and config
typer[all]>=0.9
pydantic>=2.0
pyyaml

# API server
fastapi>=0.115
uvicorn[standard]>=0.30

# Git & Git provider SDKs
GitPython>=3.1
PyGithub>=2.3
python-gitlab>=4.11

# HTTP & webhooks
httpx>=0.27
requests>=2.32

# Docker & registries
docker>=7.1
semver>=3.0

# Database & migrations
sqlalchemy>=2.0
alembic>=1.13

# Scheduling
apscheduler>=3.10

# Optional: better concurrency + utils
anyio>=4.4